Created payment screen

+++ app/src/main/java/com/menesdurak/e_ticaret_uygulamasi/common/Extensions.kt
@@ -1,6 +1,6 @@
 package com.menesdurak.e_ticaret_uygulamasi.common
 
-infix fun Float.round(decimals: Int): Float {
+infix fun Double.round(decimals: Int): Double {
     var multiplier = 1
     repeat(decimals) { multiplier *= 10 }
     return kotlin.math.round(this * multiplier) / multiplier
